chat Module renamed to siq
