--- native/zohodesk_portal_chat/README.md
@@ -1,6 +1,0 @@
-# zohodesk_portal_chat
-
-ASAP SDK chat flutter plugin
-
-
-
