--- native/zohodesk_portal_chat/android/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'zohodesk_portal_chat'
