--- native/zohodesk_portal_chat/lib/zohodesk_portal_chat.dart
@@ -1,9 +1,0 @@
-
-import 'zohodesk_portal_chat_platform_interface.dart';
-
-///ASAP Agent Chat
-class ZDPortalChat {
-
-  ///To show Agent Chat screen
-  static Future<void> show() => ZohodeskPortalChatPlatform.instance.show();
-}
