+++ native/zohodesk_portal_siq/android/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'zohodesk_portal_siq'
