+++ native/zohodesk_portal_siq/android/src/main/AndroidManifest.xml
