+++ native/zohodesk_portal_siq/android/src/main/kotlin/com/zoho/desk/asap/siq/zohodesk_portal_siq/ZohodeskPortalSiqPlugin.kt
@@ -1,18 +1,16 @@
-package com.zoho.desk.asap.chat.zohodesk_portal_chat
+package com.zoho.desk.asap.siq.zohodesk_portal_siq
 
 import android.app.Activity
 import androidx.annotation.NonNull
-import io.flutter.app.FlutterPluginRegistry
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
-import io.flutter.plugin.common.PluginRegistry
-import com.zoho.desk.chat.ZDPortalChat
+import com.zoho.desk.asap.siq.ZohoDeskPortalSalesIQ
 
-/** ZohodeskPortalChatPlugin */
-class ZohodeskPortalChatPlugin: FlutterPlugin, MethodChannel.MethodCallHandler, ActivityAware {
+/** ZohodeskPortalSiqPlugin */
+class ZohodeskPortalSiqPlugin: FlutterPlugin, MethodChannel.MethodCallHandler, ActivityAware {
   /// The MethodChannel that will the communication between Flutter and native Android
   ///
   /// This local reference serves to register the plugin with the Flutter Engine and unregister it
@@ -21,14 +19,14 @@ class ZohodeskPortalChatPlugin: FlutterPlugin, MethodChannel.MethodCallHandler, 
   protected var activity : Activity? = null
 
   override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
-    channel = MethodChannel(flutterPluginBinding.binaryMessenger, "zohodesk_portal_chat")
+    channel = MethodChannel(flutterPluginBinding.binaryMessenger, "zohodesk_portal_siq")
     channel.setMethodCallHandler(this)
   }
 
   override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: MethodChannel.Result) {
     when(call.method) {
       "show" -> activity?.let {
-        ZDPortalChat.show(activity)
+        ZohoDeskPortalSalesIQ.show(activity)
       } ?: kotlin.run { result.notImplemented() }
       else -> result.notImplemented()
     }
