+++ native/zohodesk_portal_siq/ios/zohodesk_portal_siq.podspec
@@ -1,10 +1,10 @@
 #
 # To learn more about a Podspec see http://guides.cocoapods.org/syntax/podspec.html.
-# Run `pod lib lint zohodesk_portal_chat.podspec` to validate before publishing.
+# Run `pod lib lint zohodesk_portal_siq.podspec` to validate before publishing.
 #
 Pod::Spec.new do |s|
-  s.name             = 'zohodesk_portal_chat'
-  s.version          = '2.0.1'
+  s.name             = 'zohodesk_portal_siq'
+  s.version          = '2.0.2'
   s.summary          = 'Zoho Desk Portal Chat'
   
   s.description      = 'Zoho Desk Portal Chat'
@@ -15,7 +15,7 @@ Pod::Spec.new do |s|
   s.source           = { :path => '.' }
   s.source_files = 'Classes/**/*'
   s.dependency 'Flutter'
-  s.dependency 'ZohoDeskPortalSalesIQ', '4.0.0-beta.4'
+  s.dependency 'ZohoDeskPortalSalesIQ', '4.0.0'
   s.platform = :ios, '13.0'
   s.ios.deployment_target  = '13.0'
 end
