+++ native/zohodesk_portal_siq/lib/zohodesk_portal_siq.dart
@@ -1,0 +1,9 @@
+
+import 'zohodesk_portal_siq_platform_interface.dart';
+
+///ASAP Agent Chat
+class ZDPortalSiq {
+
+  ///To show Agent Chat screen
+  static Future<void> show() => ZohodeskPortalSiqPlatform.instance.show();
+}
