+++ native/zohodesk_portal_siq/lib/zohodesk_portal_siq_method_channel.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
-import 'zohodesk_portal_chat_platform_interface.dart';
+import 'zohodesk_portal_siq_platform_interface.dart';
 
-/// An implementation of [ZohodeskPortalChatPlatform] that uses method channels.
-class MethodChannelZohodeskPortalChat extends ZohodeskPortalChatPlatform {
+/// An implementation of [ZohodeskPortalSiqPlatform] that uses method channels.
+class MethodChannelZohodeskPortalSiq extends ZohodeskPortalSiqPlatform {
   /// The method channel used to interact with the native platform.
   @visibleForTesting
-  final methodChannel = const MethodChannel('zohodesk_portal_chat');
+  final methodChannel = const MethodChannel('zohodesk_portal_siq');
 
   ///ASAP Chat show channel method
   @override
