+++ native/zohodesk_portal_siq/pubspec.yaml
@@ -1,4 +1,4 @@
-name: zohodesk_portal_chat
+name: zohodesk_portal_siq
 description: Zohodesk agent chat flutter plugin.
 version: 2.0.2
 homepage:
@@ -11,7 +11,8 @@ dependencies:
   flutter:
     sdk: flutter
   plugin_platform_interface: ^2.0.2
-  zohodesk_portal_apikit: 2.0.2
+  zohodesk_portal_apikit:
+    path: ../zohodesk_portal_apikit
 
 dev_dependencies:
   flutter_test:
@@ -22,8 +23,8 @@ flutter:
   plugin:
     platforms:
       android:
-        package: com.zoho.desk.asap.chat.zohodesk_portal_chat
-        pluginClass: ZohodeskPortalChatPlugin
+        package: com.zoho.desk.asap.siq.zohodesk_portal_siq
+        pluginClass: ZohodeskPortalSiqPlugin
       ios:
-        pluginClass: ZohodeskPortalChatPlugin
+        pluginClass: ZohodeskPortalSiqPlugin
